распилила код на модули
