+++ index.html
@@ -227,7 +227,8 @@
     </section>
   </template>
 
-  <script src="/js/main.js"></script>
+  <!-- <script src="/js/main.js"></script> -->
+  <script src="/js/main.js" type="module"></script>
 
 </body>
 </html>
