+++ js/consts.js
@@ -1,0 +1,17 @@
+// модуль содержит константы проекта, из ТЗ, например
+const PHOTOS_COUNT = 25;
+const LIKES_COUNT = {
+  min: 15,
+  max: 200,
+};
+const COMMENTS_COUNT = {
+  min: 0,
+  max: 30,
+};
+
+const AVATAR_NUMBER = {
+  min: 1,
+  max: 6,
+};
+
+export { PHOTOS_COUNT, LIKES_COUNT, COMMENTS_COUNT, AVATAR_NUMBER };
