+++ js/mocks.js
@@ -1,0 +1,26 @@
+// модуль хранит моки. Потом его заменит модуль загрузки данных с сервера
+const descriptions = [
+  'ежик в тумане',
+  'черное море, белый мерседес',
+  'еще какое-нибудь описание',
+  'надо придумать что-нибудь еще',
+  'белка на айсберге',
+  'чебурашка и крокодил Гена',
+  'кот Леопольд',
+  'Чертенок №13',
+  'домовенок Кузя',
+  'фантазия кончилась',
+];
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const authorsNames = ['Ваня', 'Аня', 'Петя', 'Коля', 'Маша', 'Катя', 'Саша'];
+
+export { descriptions, messages, authorsNames };
